Only clean resources associated with your volume
When a read-only downstairs is no longer referenced by a volume, Nexus
must issue DELETE calls to the appropriate Crucible agent to clean them
up. As part of the snapshot delete saga, Nexus will decrement reference
counts and return a list of regions and region_snapshots to delete
(along with their datasets), but currently this list of region_snapshots
will contain *any* resources to clean up, not just the argument
volume's:

    volume_1 references to RO1, RO2, RO3
    volume_2 references to RO4, RO5, RO6

    resources_to_clean(volume_1) => [RO1, RO2, RO3, RO4, RO5, RO6]
    resources_to_clean(volume_2) => [RO1, RO2, RO3, RO4, RO5, RO6]

Handing these duplicate lists of resources to clean up is mostly ok: the
Crucible agent is idempotent for equivalent requests, as it has to be to
be used by Nexus in a saga. The one case where it isn't is when issuing
a DELETE call for a read-only downstairs booted from a snapshot that
doesn't exist: it will return 404. This makes sagas sad, and will cause
them to unwind.

Fix this by only returning the resources associated with the argument
volume:

    resources_to_clean(volume_1) => [RO1, RO2, RO3]
    resources_to_clean(volume_2) => [RO4, RO5, RO6]