Add `Secure` attribute to session cookie (#3902)
We had a TODO item in the code pointing to #249, but no dedicated issue.
The `Secure` attribute means the browser will only include the cookie
with encrypted requests. This is a no-brainer table stakes kind of thing
we _must_ do. This was blocked on adding TLS to Nexus.

We use `deployment.dropshot_external.tls` to decide whether to include
`Secure` so things will still work in non-TLS environments like local
dev.

~~Because the tests do not run in a browser, I think they should be
indifferent to the `Secure` attribute, but I'll find out when CI runs.~~

~~The browser ignores this on `localhost`, so for a simple local Nexus
instance without TLS this shouldn't be a problem. However, if people are
running Nexus without TLS for other test and dev setups, that might be a
problem — the web console won't work. I think we could get around that
by looking at the config value for `deployment.dropshot_external.tls`
and only including `Secure` if it's true.~~