Fix oximeter replicated SQL (#4053)
This PR is a offshoot from
https://github.com/oxidecomputer/omicron/pull/4040. The replicated
ClickHouse test continues to be flaky, so I have extracted the fix to
the replicated SQL to this PR.

I have manually tested the SQL, and was tested by CI on the other branch
here
https://github.com/oxidecomputer/omicron/pull/4040/checks?check_run_id=16599532664

From the previous PR explaining the fix:

> There were also some DB::Exception errors in the replicated SQL
config. These would be caught by helios/deploy in practice, but since
the feature is disabled, the test passed. These errors came from the
DistributedEngine tables SQL. ORDER_BY and TTL are not supported in
DistributedEngine. Additionally, the sharding key must be a uint. While
I tested the ReplicatedEngine tables manually thoroughly, I relied on
the test for the DistributedEngine tables which were added in the last
iteration of the PR.