[wicket] Add hostname and purpose checks to uploaded SSL certs (#4086)
This is a rework of #3436; I think it's enough different that it
warrants a separate review. Many of the comments on the previous PR were
on the bits where I was implementing the checks; this version leans much
more heavily on OpenSSL to do those checks.

This addresses the initial bit of #4045 (validating names on certs prior
to RSS), but not the full bit: we should also validate names when
creating silos. That didn't look completely trivial to plumb through, so
I left a `TODO` and will try to tackle that in a followup PR.