Split regions ensure undo into separate saga node (#4055)
In both the disk_create and snapshot_create sagas, split the saga nodes
to deal with `ensure_all_datasets_and_regions` partially failing. See
RFD 419 for more information on this pattern. `delete_crucible_region`
also has to change to see if the region never existed, because the real
Crucible agent would return a 404 for a DELETE of a non-existent region.

Catch the case where we're trying to delete a Crucible region that never
existed. As well, the real Crucible agent would not let a Failed region
be deleted, so change that. This caused a test to fail, that should
never have passed! We cannot delete Failed regions and this test was
checking that we could. Correct this by causing a disk whose regions
cannot be deleted to be Faulted.

Finally, ensure that there are no stuck sagas during our tests.