Explicitly manage schema upgrade transactions (#4096)
- Previously, schema updates encouraged authors on each change to add
their own transactions, validating that the "current" and "target"
versions are correct.
- This unfortunately is not handled particularly well in scripted SQL. I
**incorrectly** thought that failing a transaction while
`batch_execute`-ing it (e.g., via a `CAST` error) would cause the
transaction to fail, and rollback. **This is not true**. In CockroachDB,
an error is thrown, but the transaction is not closed. This was the
cause of #4093 , where connections stuck in this mangled ongoing
transaction state were placed back into the connection pool.
- To fix this: Nexus now explicitly wraps each schema change in a
transaction using Diesel, which ensures that "on success, they're
committed, and on failure, they're rolled back").
- Additionally, this PR upgrades all existing schema changes to conform
to this "implied transaction from Nexus" policy, and makes it possible
to upgrade using multiple transactions in a single version change.

Fixes https://github.com/oxidecomputer/omicron/issues/4093