Define vmm table & update instance table
Create schema version 6.0.0, which defines the vmm table, updates the instance
table to refer to it, and updates the sled_instance view accordingly, and update
dbinit.sql accordingly.

Slightly reorder the tests in the schema upgrade tests so that table constraints
are checked before CHECK constraints. I found the table constraints to be a
little more actionable than the CHECK constraints (they do a better job of
describing exactly which tables have moved around in the case where tests are
failing because the order of table definitions changed).

Add tips for adding tables, modifying views, and renaming columns to the schema
README.

Tested by running the schema integration tests. Basically nothing else will work
at this point because the rest of Nexus hasn't yet learned about the new schema.