sled agent: rework instance state management
N.B. This is one of the larger functional changes on the sled agent side of this
change sequence and could use some extra attention.

Update the `InstanceStates` structure in `common/instance.rs` to track instance
and VMM runtime states separately, since these are now independent data types.
This structure also needs to track the ID of the VMM it's tracking so that it
can produce a `SledInstanceStates` on demand to send back to Nexus. (Note that
this ID can no longer be inferred from the instance runtime state--see below.)

Significantly rework the algorithm for completing live migration. Previously,
migration sources couldn't transfer control to migration targets on observing
a successful migration, because instance records only contained the target
Propolis's ID and not its sled ID or server IP. Now, having the new Propolis ID
is enough to change the instance's "VMM pointer," which is enough to change the
sled ID and Propolis IP implicitly.

Add logic to "retire" an active Propolis if sled agent observes a "destroyed" or
"failed" Propolis state and the Propolis in question is still the instance's
active VMM.

Update the unit tests to check instance and VMM states separately. Add checks
that verify that an instance's generation number changes if and only if one of
the fields protected by that generation number has changed.

This commit does not update the many callers of `InstanceStates` who now need to
adapt to its new API surface.