Fix up simulated sled agent collection tests
These need some minor restructuring to handle the new shape of the
InstanceStates structure.

Do away with the `is_stopped` function for instance states; this isn't used
anywhere outside of tests, and the tests can just set their own expectations for
whether a VMM should pass through the Stopping state or not.