Add an SMF property for rack_id to MGS (#4145)
A follow up commit will plumb the rack_id along with any discovered
sleds to a new "hardware_sled" table in CRDB used for sled addition to
an already initialized rack. This is going to be required when adding
sleds to initialized racks, in order to create a `StartSledAgentRequest`
for new sleds.