Draft CTE for updating an instance/VMM in a single statement
Sled agents now (or will soon) send Nexus updates in the form of a
`SledInstanceStates` struct that contains updated instance and VMM state in a
single package. Write a CTE that updates an instance and VMM in a single
atomic statement to allow sled agent to assume that instance and VMM updates
will be applied atomically. See the comments in the new query module for more
details.