nexus: fix network interface queries
The network interface query module has a handcrafted subquery that tries to
reason about an instance's state using only the instance record. Fix this up to
also account for whether the instance has an active VMM. Also update the tests
in this module to paste a fake VMM ID into an instance in lieu of setting its
state to Running.

Fix other portions of the network interface test build that were broken by
changes to the instance record's format.

Add a `From` impl to convert api::internal::nexus::InstanceRuntimeState to
nexus_db_model::InstanceRuntimeState. This isn't technically needed by this
change (I misread a build error) but is going to be needed soon enough.