[db] omdb db disk for physical disks (#4141)
New omdb command and database schema updates

This bumps the database version to 5.0.0
This adds a new subcommand: `physical <UUID>` to `omdb db disks`.  
This will show the disk resources that reside on a physical disk.

This change requires additional indexes in the zpool and dataset tables as
we lookup which zpool is on a physical disk, and we lookup which
datasets are on a given zpool. The usage here is just for the omdb tool,
but those same searches will need to happen when Omicron needs to figure out
what resources are on a physical disk.

Tested both initial install, and schema update
---------

Co-authored-by: Alan Hanson <alan@oxide.computer>