Update Nexus::instance_request_state
The outcome of requesting a state change now depends in part on whether an
instance has an active VMM and, if so, what that VMM's state is.

Requesting state changes also requires an instance's current VMM to determine
what sled to send state change requests to.

This commit fixes up callers of instance_request_state in instance.rs itself
but not in the sagas that call it (sagas will be updated later in their own
individual commits).