Rework instance create saga
Merely creating an instance no longer assigns a sled ID or a Propolis ID/IP,
so remove all those nodes from the create saga.

To handle create-and-start requests, make the create saga move newly-created
instances to Stopped, then have the create saga's invoker run a start saga once
the request to create the instance finishes. This means the create saga no
longer has to know how to set up instance networking state, so remove all those
nodes too. (Note that this is a minor change in behavior: previously, if a
client successfully generated a create saga and then its Nexus server crashed,
the client call would fail but the instance would still start, because that
was part of the saga; now, the instance will be created but not started. If
desired this can be addressed in the future by making the start saga a subsaga
of the create saga, but that work is not relevant to the instance/VMM split and
so is left for another day.)