Update instance start saga
This saga now has to allocate a sled for the instance to run on and give its VMM
an IP on that sled, so start doing that.

Create a new module with some common instance operations that this saga shares
with the migration saga (which will be cleaned up next).