Update instance migration saga
Change the migration saga to create VMM records and to use common code to
reserve space on the destination sled.

Splitting VMM information off from the instance table allows the source and
target sled agents to use the same "pre-migration" instance record--it's no
longer necessary to synthesize an instance record for the target with an
as-yet-unused Propolis ID, and accordingly it's no longer necessary to do extra
work to avoid writing that record back to CRDB. Moreover, sled agent now
understands whether a VMM is the "active" VMM for an instance, so it can
guarantee that unregistering a prospective migration target (as in saga unwind)
won't produce a new instance record. This considerably simplifies saga unwind.
Rewrite the saga comments accordingly.

Also take this opportunity to drop the `WriteBack` discipline enum from instance
unregistration, since it's no longer needed.

Add a sled agent unit test that explicitly verifies that rudely terminating a
migration target VMM leaves its instance record unchanged.