Update disk snapshot saga
When creating a snapshot saga for an attached disk, get its instance's current
state and sled ID (if it has one) and use these to construct an optional
instance-and-sled tuple to pass to the saga instead of setting a "use the
pantry" flag to false and then looking this information up again mid-saga.

The old and new code are morally equivalent from a synchronization standpoint:
it was and still is possible for an instance to stop or move between the time
the saga decides what sled agent to talk to and the time the snapshot request
is issued. (The window is a little wider with this change in place, though.)