bugfix: reorder instance columns in Diesel table schema
Diesel's struct derives assume that the order of fields in a struct matches
the order of columns in the corresponding `table!` directive. Make sure this
is the case for the instance table (the static machine properties moved when
they were promoted from the runtime state to the regular instance record).