bugfix: use a valid state name in instance state subquery
The relevant CASE and COALESCE statements need to return a valid instance state
in all their prongs. Represent that an instance is running if it has an active
VMM--the important thing is that it's in a state where it's not safe to modify
its network interfaces, not that it's in any *particular* state of the sort.