Don't abort instance monitor task on RunningState drop
If Propolis moves to the Destroyed state, the instance state monitor (now) calls
`terminate` on the relevant sled agent Instance before publishing the instance's
final state update to Nexus. Make sure that this doesn't abort the monitor task
so that the subsequent publishing of the last state update actually goes
through.

(This wasn't previously a problem because the instance state monitor published
the terminal state first and then terminated the zone. Changing this ordering
is intentional--it ensures that Propolis zones are gone before their resources
are released and that an instance can't be restarted until it has been
unregistered from its sled agent's instance manager.)