Remove instances from instance manager if terminated before starting
Ensure that if an instance is terminated before its Propolis zone is started,
it is removed from the instance manager's instance map so that the instance can
be restarted on that sled.

Found via omicron-stress, which is much happier with this fix in place.

Also improve a little bit of tracing.