[workspace-hack] use workspace-dotted format and a patch directive (#4197)
Two changes:

1. Switch to the workspace-dotted format (`.workspace = true`) for
   uniformity with the rest of omicron. This is new in cargo-hakari
   0.9.28.

2. Use a patch directive, which means that the workspace-hack only
   applies while building within this workspace. If another workspace
   imports a crate from here via a git dependency, it will not have the
workspace-hack applied to it (instead, it will use [this empty
crate](https://crates.io/crates/omicron-workspace-hack)
   on crates.io). Thanks so much to @pfmooney for this
   suggestion!

Also remove one of the exceptions made in the xtask (workspace-hack
lines in *other* `Cargo.toml`s are now output as `.workspace = true`,
but hakari cannot yet generate workspace lines in its own `Cargo.toml`).

I verified by creating an empty project that the workspace-hack isn't
applied to downstream projects that import e.g. `omicron-common` as a
git (or path) dependency.

Folks will have to update to cargo-hakari 0.9.28 to use this, but
hopefully that won't be too much of a bother.