Adds functionality to run oximeter standalone (#4117)
- Adds a "standalone" mode for the `oximeter-collector` crate, including
the binary and main inner types. This runs in a slightly different mode,
in which the ClickHouse database itself isn't strictly required. In this
case, a task to simply print the results will be spawned in place of the
normal results-sink task which inserts records into the database.
- Creates a tiny fake Nexus server, which includes only the API needed
to register collectors and producers. This is started automatically when
running `oximeter standalone`, and used to assign producers / collectors
as the real Nexus does, but without a database. The assignments are only
in memory.
- Adds internal `oximeter` API for listing / deleting a producer for
each oximeter collector, and an `omdb` subcommand which exercises the
listing.