make Nexus sole owner of Instance records' state values
Sled agent has one path that updates the `state` column in an instance record
(the `fallback_state` field in an `InstanceRuntimeState`): if an active Propolis
is retired (i.e. the instance's Propolis ID goes from `Some` to `None`), the
instance goes to the Stopped state.

This rule is so simple that it can be implemented entirely in Nexus, giving
Nexus sole ownership of this table column:

- Remove the `fallback_state` field from `api::internal::InstanceRuntimeState`.
  This completely hides the field from the sled agent.
- When Nexus processes a new `SledInstanceState`, check the
  `active_propolis_id`: if it's `Some`, the state is Running; if it's `None`,
  the state is Stopped.

The only other states that an instance can now have are Creating and Destroyed,
both of which are set only by Nexus and only when it is known that no sled agent
refers to the instance (either by definition when an instance is being created,
or because the instance has no active VMM when it's being destroyed).

Rename the `fallback_state` field to `nexus_state` to reflect its new ownership.