Move instance/VMM table schema upgrade to version 7.0.0
The instance/VMM table schema change was slated to be 6.0.0 in the original
version of its pull request. That version was then added by a separate PR, but
this didn't cause a merge conflict because the instance/VMM upgrade used an
extra "0" in its schema upgrade files (for fear that there might be more than
nine of them and that the leading 0 would be necessary to ensure they had the
correct lexographical ordering).

The schema changes don't conflict with each other, so everything actually works
fine, but having two logically separate updates in one version is at the very
least aesthetically displeasing.

Move the instance schema upgrade to version 7.0.0. Rename the files to remove
the leading 0 in their numbers, since that turned out not to be needed.

Tested via cargo tests (there are no other functional or schema changes beyond
renaming and updating version constants).